Fix handling of more than 1 new pull request in a single poll
We handle pull requests in descending order of ID. Currently, as soon as we
post a comment successfully, we update status['pull_req_number'], which
means that upon seeing the next pull request, we immediately bail as the ID
is now under the threshold.

Save the value of status['pull_req_number'] at the beginning of the poll to
avoid this.

Signed-off-by: Andrew Donnellan <andrew@donnellan.id.au>