Added a comment condition
The bot doesn't check if the pull request it attempts to comment on is
closed or not. Instances of `PullRequest` have a `closed_at` variable,
which will return either `None` or a datetime, if the request hasn't or
has been closed when the bot checks the PR.

This prevents the bot from unnecessarily attempting to comment on or close
an already closed pull request, which may also speed up the execution, as
it won't make API calls for closed PRs.

Signed-off-by: Grayson "Sweets"
[ajd: fix whitespace]
Signed-off-by: Andrew Donnellan <andrew@donnellan.id.au>